[auth.ts, page.tsx] Updated
V1.1
- [ x ] Authentication
- [ x ] Accessing User's data
- [ x ] Home Page and Main Page
**IN PROGRESS**
- [ ] Storage System (GAIA)
