+++ frontend/src/app/newUser/page.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function newUser() {
+  return (
+    <div>
+      Hello
+    </div>
+  )
+}
